feat: add solving to module folder
